+++ packages/orq-rc/RELEASES.md
@@ -8859,3 +8859,13 @@ Based on:
 - [typescript v4.9.0-rc.24] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.24 - packages/orq-rc
+
+## 2026-05-06 04:32:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.25] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.25 - packages/orq-rc